requirements.txt
numpy 
jax-metal #==0.1.0 #for jax on mac 
jax 
jaxlib #>=v0.4.26
flax
optax
tensorboard
torchvision
matplotlib
hydra-core
orbax-export